# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks the caller can see, with optional filters
      parameters:
        - { name: assignee, in: query, required: false, schema: { type: string, enum: [me, anyone] } }
        - { name: status, in: query, required: false, schema: { type: string, enum: [open, in_progress, completed, cancelled] } }
        - { name: patientId, in: query, required: false, schema: { type: string } }
        - { name: encounterId, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskListResponse"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a clinical task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Patient or encounter not found
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task by id
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Edit a task (title, due date, priority, assignee, status)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
  /tasks/{id}/complete:
    post:
      operationId: completeTask
      tags: [tasks]
      summary: Mark a task complete
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Task completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
        "409":
          description: Task already closed
  /tasks/{id}/cancel:
    post:
      operationId: cancelTask
      tags: [tasks]
      summary: Cancel a task
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelTaskRequest"
      responses:
        "200":
          description: Task cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
        "409":
          description: Task already closed
  /encounters/{id}/tasks/generate:
    post:
      operationId: generateEncounterTasks
      tags: [tasks]
      summary: Run AI task generation on the encounter's note
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskGenerateResponse"
        "404":
          description: Encounter not found
        "409":
          description: Encounter has no associated note yet
...
components:
  schemas:
...
    TaskCategory:
      type: string
      enum:
        - call_patient
        - schedule_followup
        - send_referral
        - prior_auth
        - obtain_records
        - repeat_labs
        - nursing_instruction
        - billing_followup
        - patient_instruction
        - other
    TaskStatus:
      type: string
      enum: [open, in_progress, completed, cancelled]
    TaskPriority:
      type: string
      enum: [low, normal, high]
    Task:
      type: object
      required:
        - id
        - encounterId
        - category
        - title
        - description
        - dueAt
        - priority
        - status
        - isClosed
        - source
      properties:
        id: { type: string }
        encounterId: { type: ["string", "null"] }
        category: { $ref: "#/components/schemas/TaskCategory" }
        title: { type: string }
        description: { type: ["string", "null"] }
        dueAt:
          type: ["string", "null"]
          format: date-time
        priority: { $ref: "#/components/schemas/TaskPriority" }
        status: { $ref: "#/components/schemas/TaskStatus" }
        isClosed: { type: boolean }
        source:
          type: string
          enum: [ai, manual]
    TaskListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Task" }
    TaskGenerateResponse:
      type: object
      required: [data, source]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Task" }
        source:
          type: string
          enum: [ai, stub]
    CreateTaskRequest:
      type: object
      required: [patientId, title, category]
      properties:
        patientId: { type: string }
        encounterId: { type: string }
        title:
          type: string
          minLength: 1
          maxLength: 200
        description: { type: string, maxLength: 2000 }
        category: { $ref: "#/components/schemas/TaskCategory" }
        priority: { $ref: "#/components/schemas/TaskPriority" }
        dueAt:
          type: string
          format: date-time
        assignedUserId: { type: string }
    UpdateTaskRequest:
      type: object
      description: Partial update — only fields present in the body are touched.
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description: { type: ["string", "null"], maxLength: 2000 }
        category: { $ref: "#/components/schemas/TaskCategory" }
        priority: { $ref: "#/components/schemas/TaskPriority" }
        dueAt:
          type: ["string", "null"]
          format: date-time
        assignedUserId:
          type: ["string", "null"]
        status: { $ref: "#/components/schemas/TaskStatus" }
    CancelTaskRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 500
...